examples/google_api/requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dateutil>=2.8.0
tqdm>=4.65.0 
python-dotenv>=1.1.0
bs4==0.0.2
sqlmodel>=0.0.24
chonkie>=1.0.0
slack-sdk>=3.21.0
